requirements.txt
flet
pillow
pymupdf
easyocr
g4f
langdetect
ebooklib
reportlab